Ajustes

+++ ConversorDolar/Models/ExchangeRateResult.cs
@@ -6,7 +6,7 @@
         public string? License { get; set; } = string.Empty;
         public long Timestamp { get; set; }
         public string? Base { get; set; } = string.Empty;
-        public System.Collections.Generic.Dictionary<string, decimal>? Rates { get; set; }
+        public System.Collections.Generic.Dictionary<string, double>? Rates { get; set; }
 
     }
 }
